Fix RedRatHub error control function
Fix an issue where more than 1 RedRatHub process could be started when managing errors.
Introduce a file lock so only a single stbControl.pl process can handle the RedRatHub
restart process.

+++ scripts/stbControl.pl
@@ -2,7 +2,7 @@
 
 use strict;
 use JSON;
-use Fcntl;
+use Fcntl ':flock';
 use CGI;
 use Tie::File::AsHash;
 use Time::HiRes qw (sleep);
@@ -57,8 +57,10 @@ if (-e $stbdatafile) {
         local $/ = undef;
         open my $fh, "<", $stbdatafile or die "ERROR: Unable to open $stbdatafile: $!\n";
         my $data = <$fh>;
-        my $decoded = $json->decode($data);
-        %stbdata = %{$decoded};
+	if ($data) {
+	        my $decoded = $json->decode($data);
+        	%stbdata = %{$decoded};
+	}
 }
 
 ##### Load the groups data
@@ -67,8 +69,10 @@ if (-e $groupsfile) {
         local $/ = undef;
         open my $fh, "<", $groupsfile or die "ERROR: Unable to open $groupsfile: $!\n";
         my $data = <$fh>;
-        my $decoded = $json->decode($data);
-        %groups = %{$decoded};
+	if ($data) {
+	        my $decoded = $json->decode($data);
+        	%groups = %{$decoded};
+	}
 }
 
 ##### Load the sequences data
@@ -77,8 +81,10 @@ if (-e $seqfile) {
         local $/ = undef;
         open my $fh, "<", $seqfile or die "ERROR: Unable to open $seqfile: $!\n";
         my $data = <$fh>;
-        my $decoded = $json->decode($data);
-        %seqs = %{$decoded};
+	if ($data) {
+	        my $decoded = $json->decode($data);
+        	%seqs = %{$decoded};
+	}
 }
 
 chomp(my $action = $ARGV[0] // $query->param('action') // '');
@@ -657,18 +663,6 @@ sub checkRedRatHub {
 		die "CRITICAL ERROR: Unable to identify host architecture for checking RedRatHub process in stbControl.pl\n";
 	}
 
-	#my $processidfile = $filedir . 'redRadHubID.txt';
-	#my $redrathubid = '';
-	#if (-e $processidfile) {
-	#	if (open my $fh, '<', $processidfile) {
-	#		local $/;
-	#		my $info = <$fh>;
-	#		if () {
-
-	#		}
-	#	}
-	#}
-
 	chomp(my $running = `ps -ax | grep "stbController-RedRatHubProcess" | grep -v grep` // '');
         if (!$running) {
                 print "RedRatHub which controls IRNetBoxIV communication is NOT running. I have tried to start it for you. Please wait 10 seconds and then try to control IRNetBoxIV devices again\n\nIf the problem persists, check your webserver error log for details or contact your system administrator.";
@@ -676,7 +670,14 @@ sub checkRedRatHub {
                 $sysip =~ s/\s+//g;
                 $sysip =~ s/\r|\n//g;
                 if ($sysip) {
-			system("cd $redrathubdir && bash -c \"exec -a stbController-RedRatHubProcess-$sysip $dotnetbin RedRatHub.dll --noscan --nohttp > $redrathubdebug 2>&1 \&\" \&");
+			my $lfh;
+			my $lockfile = $filedir . 'redRatHub.lock';
+			open $lfh, '+>', $lockfile or print "ERROR: Unable to open $lockfile for writing and locking: $!\n" and return;
+			if (flock($lfh, LOCK_EX | LOCK_NB)) {
+				system("cd $redrathubdir && bash -c \"exec -a stbController-RedRatHubProcess-$sysip $dotnetbin RedRatHub.dll --noscan --nohttp > $redrathubdebug 2>&1 \&\" \&");
+				flock($lfh, LOCK_UN);
+				close $lfh;
+			}
 			return \"fail";
                 } else {
                         die "Could not identify the system ip address for the RedRatHub process. STB controller IR requires this.\n";
@@ -687,33 +688,40 @@ sub checkRedRatHub {
         	##### If the RedRatHub is running, check to see if there are errors in the log file which suggest a restart would help
 		my $errors = `grep -i "exception\\\|canceled" $redrathubdebug` // '';
 		if ($errors) {
-			if (!$logpid) {
-				print "Multiple errors have been detected with the RedRatHub process that supports RedRat IR. I have restarted the process for you. If this problem persists, please contact your system administrator for assistance.";
-			}
-			##### Kill the current RedRatHub process
-			system("kill $runpid");
+			my $lfh;
+			my $lockfile = $filedir . 'redRatHub.lock';
+			open $lfh, '+>', $lockfile or print "ERROR: Unable to open $lockfile for writing and locking: $!\n" and return;
+			if (flock($lfh, LOCK_EX | LOCK_NB)) {
+				if (!$logpid) {
+					print "Multiple errors have been detected with the RedRatHub process that supports RedRat IR. I have restarted the process for you. If this problem persists, please contact your system administrator for assistance.";
+				}
+				##### Kill the current RedRatHub process
+				system("kill $runpid");
 
-			##### Clear out the current debug log file
-			if (open my $fh, '+>',$redrathubdebug) {
-				close $fh;
-			} else {
-				warn "Unable to overwrite the file $redrathubdebug for error reset. $!\n";
-			}
-			
-			##### Start the new RedRatHub process
-	                chomp(my $sysip = `hostname -I | awk \'\{print \$1\}\'` // '');	# Get the systems IP address
-	                $sysip =~ s/\s+//g;
-	                $sysip =~ s/\r|\n//g;
-	                if ($sysip) {
-				system("cd $redrathubdir && bash -c \"exec -a stbController-RedRatHubProcess-$sysip $dotnetbin RedRatHub.dll --noscan --nohttp > $redrathubdebug 2>&1 \&\" \&");
-	                } else {
-	                        die "Could not identify the system ip address for the RedRatHub process. STB controller IR requires this.\n";
-	                }
-	                
-			if ($logpid) {	# If this script has been run from the event scheduler, wait for it to restart before carrying on
-				sleep 5;
-			} else {	# Otherwise this has come from the front end control so just stop and return a fail status
-				return \"fail";
+				##### Clear out the current debug log file
+				if (open my $fh, '+>',$redrathubdebug) {
+					close $fh;
+				} else {
+					warn "Unable to overwrite the file $redrathubdebug for error reset. $!\n";
+				}
+				
+				##### Start the new RedRatHub process
+		                chomp(my $sysip = `hostname -I | awk \'\{print \$1\}\'` // '');	# Get the systems IP address
+		                $sysip =~ s/\s+//g;
+		                $sysip =~ s/\r|\n//g;
+		                if ($sysip) {
+					system("cd $redrathubdir && bash -c \"exec -a stbController-RedRatHubProcess-$sysip $dotnetbin RedRatHub.dll --noscan --nohttp > $redrathubdebug 2>&1 \&\" \&");
+		                } else {
+		                        die "Could not identify the system ip address for the RedRatHub process. STB controller IR requires this.\n";
+		                }
+		                
+				if ($logpid) {	# If this script has been run from the event scheduler, wait for it to restart before carrying on
+					sleep 5;
+				} else {	# Otherwise this has come from the front end control so just stop and return a fail status
+					return \"fail";
+				}
+				flock($lfh, LOCK_UN);
+				close $lfh;
 			}
 		}
